refactor: move RangeSlider in its own file
